requirements.txt
celery==5.1.2
dj-database-url==0.5.0
Django==3.2.6
django-celery-results==2.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
